@@ -1,10 +1,12 @@
 {
   "name": "rxjs-in-ng",
   "version": "0.2.0",
-  "license": "MIT",
   "scripts": {
+    "demo": "npm-run-all -p demo:*",
+    "demo:serve": "yarn start",
+    "demo:app": "yarn wait-on http-get://localhost:8806 && google-chrome --app=http://localhost:8806",
     "ng": "ng",
-    "start": "ng serve --port 8806 --host nix",
+    "start": "ng serve --port 8806 ",
     "build": "ng build --prod",
     "test": "ng test",
     "lint": "ng lint",
@@ -12,38 +14,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "angular-in-memory-web-api": "^0.5.3",
-    "core-js": "^2.4.1",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
-  },
-  "devDependencies": {
-    "@angular/cli": "1.7.3",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.5.5",
+    "@angular-devkit/build-ng-packagr": "~0.5.0",
+    "@angular/animations": "^6.0.0-rc.3",
+    "@angular/cli": "6.0.0-rc.4",
+    "@angular/common": "^6.0.0-rc.3",
+    "@angular/compiler": "^6.0.0-rc.3",
+    "@angular/compiler-cli": "^6.0.0-rc.3",
+    "@angular/core": "^6.0.0-rc.3",
+    "@angular/forms": "^6.0.0-rc.3",
+    "@angular/http": "^6.0.0-rc.3",
+    "@angular/language-service": "^6.0.0-rc.3",
+    "@angular/platform-browser": "^6.0.0-rc.3",
+    "@angular/platform-browser-dynamic": "^6.0.0-rc.3",
+    "@angular/router": "^6.0.0-rc.3",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
+    "@types/node": "^9.6.4",
+    "angular-in-memory-web-api": "0.6.0",
     "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "core-js": "^2.5.5",
+    "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~2.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
+    "karma-coverage-istanbul-reporter": "^1.4.2",
+    "karma-jasmine": "~1.1.1",
+    "karma-jasmine-html-reporter": "^1.0.0",
+    "ng-packagr": "^2.4.1",
+    "npm-run-all": "^4.1.2",
+    "protractor": "^5.3.1",
+    "rxjs": "^6.0.0-uncanny-rc.7",
+    "ts-node": "^5.0.1",
+    "tsickle": ">=0.25.5",
+    "tslib": "^1.9.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
-  }
+    "typescript": "~2.7.2",
+    "wait-on": "^2.1.0",
+    "zone.js": "^0.8.26"
+  },
+  "license": "(MIT OR Apache-2.0)",
+  "licenses": [
+    {
+      "license": "(MIT OR Apache-2.0)"
+    }
+  ]
 }
