@@ -1,51 +1,50 @@
 // tslint:disable:member-ordering
-import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Movie, People, RootMovies, RootPeople } from '../sw-interfaces';
-import { SwUrlService } from 'app/samples/sw-url.service';
-
-import { Observable } from 'rxjs/Observable';
-import { of } from 'rxjs/observable/of';
+import { Injectable } from '@angular/core';
+import { of } from 'rxjs';
 import { catchError, map } from 'rxjs/operators';
+import { RootMovies, RootPeople } from '../sw-interfaces';
+import { SwUrlService } from '../sw-url.service';
+
+
 
 @Injectable()
 export class FilmPeopleService {
-
   constructor(private http: HttpClient, private swUrlService: SwUrlService) {}
 
   getFilmsByName(title: string) {
-    return this.http.get<RootMovies>(this.movieUrl + `/?title=${title}`)
-      .pipe(
+    return this.http.get<RootMovies>(this.movieUrl + `/?title=${title}`).pipe(
+      // RxJS operators
+      map((data: RootMovies) => data.results),
 
-        // RxJS operators
-        map((data: RootMovies) => data.results),
-
-        catchError(err => {
-          // log HTTP error and ...
-          console.error('GET movies failed', err);
-          // rethrow as a user-friendly message
-          throw new Error('Sorry but can\'t get movies right now; please try again later');
-        })
+      catchError(err => {
+        // log HTTP error and ...
+        console.error('GET movies failed', err);
+        // rethrow as a user-friendly message
+        throw new Error(
+          `Sorry but can\'t get movies right now; please try again later`
+        );
+      })
     );
   }
 
   getSwPeopleByUrl(url: string) {
     const id = url ? +url.match(/\/([0-9]*)\/$/)[1] : NaN;
     if (isNaN(id)) {
-      return of([])
+      return of([]);
     }
-    return this.http.get<RootPeople>(`api/swPeople/${id}`)
-      .pipe(
+    return this.http.get<RootPeople>(`api/swPeople/${id}`).pipe(
+      // RxJS operators
+      map((data: RootPeople) => data.results),
 
-        // RxJS operators
-        map((data: RootPeople) => data.results),
-
-        catchError(err => {
-          // log HTTP error and ...
-          console.error('GET movies failed', err);
-          // rethrow as a user-friendly message
-          throw new Error('Sorry but can\'t get movies right now; please try again later');
-        })
+      catchError(err => {
+        // log HTTP error and ...
+        console.error('GET movies failed', err);
+        // rethrow as a user-friendly message
+        throw new Error(
+          'Sorry but can\'t get movies right now; please try again later'
+        );
+      })
     );
   }
   private get movieUrl() {
