@@ -1,16 +1,15 @@
-import { Component, OnInit, OnDestroy } from '@angular/core';
-import { Router, ActivatedRoute, ParamMap } from '@angular/router';
+import { Component, OnDestroy, OnInit } from '@angular/core';
+import { ActivatedRoute, Router } from '@angular/router';
 
-import { Movie } from 'app/samples/sw-interfaces';
-import { FilmService } from './film.service';
-
-import { Observable } from 'rxjs/Observable';
-import { Subscription } from 'rxjs/Subscription';
+import { Observable, Subscription } from 'rxjs';
 import { switchMap, tap } from 'rxjs/operators';
 
+import { Movie } from '../../samples/sw-interfaces';
+import { FilmService } from './film.service';
+
 @Component({
-    selector: 'app-movie',
-    template: `
+  selector: 'app-movie',
+  template: `
       <h1>Routed Movie Component</h1>
 
       <div>
@@ -27,57 +26,53 @@ import { switchMap, tap } from 'rxjs/operators';
         <button (click)="nextMovie()">Next</button>
       </div>
     `,
-    providers: [ FilmService ],
-    styleUrls: ['./movie.component.css']
+  providers: [FilmService],
+  styleUrls: ['./movie.component.css']
 })
 export class MovieComponent implements OnInit, OnDestroy {
-    private currentId: number;
-    public currentMovie: Observable<Movie>;
-    private routerEventsSubscription: Subscription;
+  private currentId: number;
+  public currentMovie: Observable<Movie>;
+  private routerEventsSubscription: Subscription;
 
-    constructor(
-        private filmService: FilmService,
-        private route: ActivatedRoute,
-        private router: Router
-    ) {
-        // Listen to the router do its thing
-        // What is `routerEventsSubscription`?
-        this.routerEventsSubscription = router.events.subscribe(event => {
-            console.log('Router event: ', event);
-        })
-    }
+  constructor(private filmService: FilmService, private route: ActivatedRoute, private router: Router) {
+    // Listen to the router do its thing
+    // What is `routerEventsSubscription`?
+    this.routerEventsSubscription = router.events.subscribe(event => {
+      console.log('Router event: ', event);
+    });
+  }
 
-    ngOnInit(): void {
-        console.log('MovieComponent initialized');
+  ngOnInit(): void {
+    console.log('MovieComponent initialized');
 
-        // Get new movie whenever the URL's id changes
-        // Question: How can the id change?
-        this.currentMovie = this.route.paramMap.pipe(
-            switchMap(params => {
-              const id = +params.get('id');
-              return this.filmService.getFilm(id).pipe(
-                // Question: how could movie be undefined?
-                tap(movie => this.currentId = movie && movie.id)
-              )
-            })
+    // Get new movie whenever the URL's id changes
+    // Question: How can the id change?
+    this.currentMovie = this.route.paramMap.pipe(
+      switchMap(params => {
+        const id = +params.get('id');
+        return this.filmService.getFilm(id).pipe(
+          // Question: how could movie be undefined?
+          tap(movie => (this.currentId = movie && movie.id))
         );
-    }
+      })
+    );
+  }
 
-    ngOnDestroy(): void {
-        console.log('MovieComponent destroyed');
+  ngOnDestroy(): void {
+    console.log('MovieComponent destroyed');
 
-        // Question: Why must we unsubscribe from the router?
-        // Question: Why do we NOT unsubscribe from `route.paramMap`
-        this.routerEventsSubscription.unsubscribe();
-    }
+    // Question: Why must we unsubscribe from the router?
+    // Question: Why do we NOT unsubscribe from `route.paramMap`
+    this.routerEventsSubscription.unsubscribe();
+  }
 
-    nextMovie() {
-        this.currentId = this.currentId ? this.currentId + 1 : 1;
-        this.router.navigate([`/movie/${this.currentId}`]);
-    }
+  nextMovie() {
+    this.currentId = this.currentId ? this.currentId + 1 : 1;
+    this.router.navigate([`/movie/${this.currentId}`]);
+  }
 
-    previousMovie() {
-        this.currentId = this.currentId ? this.currentId - 1 : 1;
-        this.router.navigate([`/movie/${this.currentId}`]);
-    }
+  previousMovie() {
+    this.currentId = this.currentId ? this.currentId - 1 : 1;
+    this.router.navigate([`/movie/${this.currentId}`]);
+  }
 }
