@@ -1,8 +1,10 @@
-import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { RootMovies, Movie } from '../sw-interfaces';
-import { SwUrlService } from 'app/samples/sw-url.service';
-import { Observable } from 'rxjs/Observable';
+import { Injectable } from '@angular/core';
+
+import { Observable } from 'rxjs';
+
+import { RootMovies } from '../sw-interfaces';
+import { SwUrlService } from '../../samples/sw-url.service';
 
 @Injectable()
 export class SimpleFilmsService {
