@@ -4,9 +4,9 @@
  *  AsyncPipe to display films
  */
 import { Component, OnInit } from '@angular/core';
-import { Movie } from 'app/samples/sw-interfaces';
+import { Movie } from '../../samples/sw-interfaces';
 
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 // Import RxJS operators
 import { catchError } from 'rxjs/operators';
