@@ -1,9 +1,9 @@
 // tslint:disable:member-ordering
 import { Component } from '@angular/core';
-import { WikipediaService, WikiResult } from 'app/samples/wikipedia.service';
+import { WikipediaService, WikiResult } from '../../samples/wikipedia.service';
 import { FormControl } from '@angular/forms';
 
-import { of } from 'rxjs/observable/of';
+import { of } from 'rxjs';
 import {
   catchError,
   debounceTime,
